Auto-detect project repo URLs
